+++ C++ League Detection Source/Detection/Objects/Tower.h
@@ -13,5 +13,5 @@ public:
     Tower();
     Position topLeft, topRight, bottomLeft, bottomRight, towerCenter;
     bool detectedTopLeft, detectedBottomLeft, detectedTopRight, detectedBottomRight;
-    float health;
+    double health;
 };
